Use RasterPixmapData until we fix QEglGLPixmapData properly
This patch includes both a change to use RasterPixmapData instead of
our own QEglRasterPixmapData as well as a fix to make QEglGLPixmapData
work to begin with. Due to the slow performance of the latter, we
should refrain from using the Egl version until we can either make webkit
not thrash pixmaps so much or improve the copying performance of egl
pixmaps.